+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - Make `FileSystemCache` crate internal. This should be used via `CosmCache`.
 - Fix return type of `FileSystemCache::load` to `VmResult<Option<Module>>` in
   order to differentiate missing files from errors.
+- Add in-memory caching for recently used Wasm modules.
 
 ## 0.11.2 (2020-10-26)